+++ src/plugins/three/images/ImageOverlayPlugin.d.ts
@@ -6,16 +6,21 @@ export class ImageOverlayPlugin {
 		overlays: Array<ImageOverlay>,
 		renderer: WebGLRenderer,
 		resolution?: number,
-		frame?: Matrix4 | null,
 	} );
 
 	addOverlay( overlay: ImageOverlay, order?: number ): void;
 	setOverlayOrder( overlay: ImageOverlay, order?: number ): void;
-	deleteOverlay( overlay: ImageOverlay, order?: number ): void;
+	deleteOverlay( overlay: ImageOverlay ): void;
 
 }
 
-export class ImageOverlay {}
+export class ImageOverlay {
+
+	color: number | Color;
+	opacity: number;
+	frame?: Matrix4 | null;
+
+}
 
 export class XYZTilesOverlay extends ImageOverlay {
 
@@ -26,6 +31,7 @@ export class XYZTilesOverlay extends ImageOverlay {
 
 		color: number | Color,
 		opacity: number,
+		frame?: Matrix4 | null,
 	} );
 
 }
@@ -37,6 +43,7 @@ export class TMSTilesOverlay extends ImageOverlay {
 
 		color: number | Color,
 		opacity: number,
+		frame?: Matrix4 | null,
 	} );
 
 }
@@ -46,10 +53,26 @@ export class CesiumIonOverlay extends ImageOverlay {
 	constructor( options: {
 		assetId: number | string,
 		apiToken: string,
-		autoRefreshToken?: boolean
+		autoRefreshToken?: boolean,
 
 		color: number | Color,
 		opacity: number,
+		frame?: Matrix4 | null,
+	} );
+
+}
+
+export class GoogleMapsOverlay extends ImageOverlay {
+
+	constructor( options: {
+		apiToken: string,
+		autoRefreshToken?: boolean,
+		logoUrl?: string,
+		sessionOptions?: null | { mapType: string, language: string, region: string },
+
+		color: number | Color,
+		opacity: number,
+		frame?: Matrix4 | null,
 	} );
 
 }